+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * Browse Lists | Hyperlink one column to improve accessibility. Refs UIIN-2266.
 * Rename `instances1` resource on `ItemsRoute` to avoid colliding with holding records. Refs UIIN-2289.
 * Z39.50 Settings toasts have a typo. Refs UIIN-2256
+* ISRI: Update the Settings create/edit screen to remove duplicated labels. Refs UIIN-2297
 
 ## [9.2.0](https://github.com/folio-org/ui-inventory/tree/v9.2.0) (2022-10-27)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v9.1.0...v9.2.0)